Maze Explorer Prototype 1
This is the first prototype to my final game Maze Explorer. In this
prototype, I've testing out the walking code for the character and the
raycasting code for the character to aim and fire. It is still
incomplete. Please not that the basic shape and level layout isn't fully
completed and will be changed in the final version. Thank you.